fix: Topbar 내부에 sidebar 컴포넌트 추가
